perf(a11y): respect prefers-reduced-motion and lazy-load Framer Motion
